pyproject.toml
[project]
name = "terraprint3d"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.0",
    "googlemaps>=4.10.0",
    "lxml>=4.9.0",
    "matplotlib>=3.8.0",
    "networkx>=3.0",
    "numpy>=2.2.6",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scipy>=1.11.0",
    "trimesh>=4.8.3",
]
